machine_learning_jaydeep/ReportAPI/requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
bleach==4.1.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.2
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastapi==0.75.1
filelock==3.6.0
flit_core==3.7.1
fonttools==4.25.0
h11==0.13.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.10.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scikit-multilearn==0.2.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.6.0
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.8.1
zipp==3.7.0
